fix(gm): address CodeRabbit review on #518
- Split oversized test modules at their seams (500-line soft cap): extract
  gm.rs's tests to gm_tests.rs via `#[path]` (542→326), and the GM-tail
  router tests to dispatch/gm_dispatch_tests.rs (tests.rs 617→469). No test
  logic changed; all 29 moved tests still run.
- Fix contradictory "Adding the next GM method" guidance: a new SGWGmPlayer
  method (index >= 109) needs no `matches!` entry — the range rule gates the
  whole tail; only 0-108 methods are named explicitly.
- Correct the stale "~85 methods" count in the gating doc (it's 117, 109-225).
- Strip PR/issue numbers from the three flagged source comments per
  copilot-instructions; keep the CAT-N-04 finding id in docs.
- Bump known-issues.md `last_updated` for the KI-14 resolution.
