+++ crates/services/src/cell/dispatch/mod.rs
@@ -30,6 +30,8 @@ mod names;
 mod router;
 
 #[cfg(test)]
+mod gm_dispatch_tests;
+#[cfg(test)]
 mod tests;
 
 // Re-export the public API at the module root so external paths
